requirements.txt
streamlit==1.39.0
langchain==0.2.10
langchain-openai==0.1.7
langchain-community==0.2.10
openai==1.40.0
faiss-cpu==1.7.4
pypdf==4.3.0
python-dotenv==1.0.1
tiktoken==0.7.0
